Construction Helpers

These functions help construct calls to various types of models.

Usage

construct_model(x, ...)

# S3 method for data.frame
construct_model(
  x,
  formula,
  method,
  method.args = list(),
  package = "base",
  env = caller_env(),
  ...
)

# S3 method for survey.design
construct_model(
  x,
  formula,
  method,
  method.args = list(),
  package = "survey",
  env = caller_env(),
  ...
)

reformulate2(
  termlabels,
  response = NULL,
  intercept = TRUE,
  pattern_term = "[ \n\r]",
  pattern_response = "[ \n\r]",
  env = parent.frame()
)

bt(x, pattern = "[ \n\r]")

bt_strip(x)

Arguments

x

construct_model.data.frame() (data.frame) a data frame

construct_model.survey.design() (survey.design) a survey design object

bt()/bt_strip() (character) character vector, typically of variable names

...

These dots are for future extensions and must be empty.

formula

(formula)
a formula

method

(string)
string naming the function to be called, e.g. "glm". If function belongs to a library that is not attached, the package name must be specified in the package argument.

method.args

(named list)
named list of arguments that will be passed to fn.

package

(string)
string of package name that will be temporarily loaded when function specified in method is executed.

env

The environment in which to evaluate expr. This environment is not applicable for quosures because they have their own environments.

termlabels

character vector giving the right-hand side of a model formula. Cannot be zero-length.

response

character string, symbol or call giving the left-hand side of a model formula, or NULL.

intercept

logical: should the formula have an intercept?

pattern_term, pattern_response

passed to bt(pattern) for arguments stats::reformulate(termlabels, response).

pattern

(string)
regular expression string. If the regex matches, backticks are added to the string. When NULL, backticks are not added.

Value

depends on the calling function

Details

construct_model(): Builds models of the form method(data = data, formula = formula, method.args!!!). If the package argument is specified, that package is temporarily attached when the model is evaluated.

reformulate2(): This is a copy of reformulate() except that variable names that contain a space are wrapped in backticks.

bt(): Adds backticks to a character vector.

bt_strip(): Removes backticks from a string if it begins and ends with a backtick.

Examples

construct_model(
  x = mtcars,
  formula = am ~ mpg + (1 | vs),
  method = "glmer",
  method.args = list(family = binomial),
  package = "lme4"
)
#> Generalized linear mixed model fit by maximum likelihood (Laplace
#>   Approximation) [glmerMod]
#>  Family: binomial  ( logit )
#> Formula: am ~ mpg + (1 | vs)
#>    Data: structure(list(mpg = c(21, 21, 22.8, 21.4, 18.7, 18.1, 14.3,  
#> 24.4, 22.8, 19.2, 17.8, 16.4, 17.3, 15.2, 10.4, 10.4, 14.7, 32.4,  
#> 30.4, 33.9, 21.5, 15.5, 15.2, 13.3, 19.2, 27.3, 26, 30.4, 15.8,  
#> 19.7, 15, 21.4), cyl = c(6, 6, 4, 6, 8, 6, 8, 4, 4, 6, 6, 8,  
#> 8, 8, 8, 8, 8, 4, 4, 4, 4, 8, 8, 8, 8, 4, 4, 4, 8, 6, 8, 4),  
#>     disp = c(160, 160, 108, 258, 360, 225, 360, 146.7, 140.8,  
#>     167.6, 167.6, 275.8, 275.8, 275.8, 472, 460, 440, 78.7, 75.7,  
#>     71.1, 120.1, 318, 304, 350, 400, 79, 120.3, 95.1, 351, 145,  
#>     301, 121), hp = c(110, 110, 93, 110, 175, 105, 245, 62, 95,  
#>     123, 123, 180, 180, 180, 205, 215, 230, 66, 52, 65, 97, 150,  
#>     150, 245, 175, 66, 91, 113, 264, 175, 335, 109), drat = c(3.9,  
#>     3.9, 3.85, 3.08, 3.15, 2.76, 3.21, 3.69, 3.92, 3.92, 3.92,  
#>     3.07, 3.07, 3.07, 2.93, 3, 3.23, 4.08, 4.93, 4.22, 3.7, 2.76,  
#>     3.15, 3.73, 3.08, 4.08, 4.43, 3.77, 4.22, 3.62, 3.54, 4.11 
#>     ), wt = c(2.62, 2.875, 2.32, 3.215, 3.44, 3.46, 3.57, 3.19,  
#>     3.15, 3.44, 3.44, 4.07, 3.73, 3.78, 5.25, 5.424, 5.345, 2.2,  
#>     1.615, 1.835, 2.465, 3.52, 3.435, 3.84, 3.845, 1.935, 2.14,  
#>     1.513, 3.17, 2.77, 3.57, 2.78), qsec = c(16.46, 17.02, 18.61,  
#>     19.44, 17.02, 20.22, 15.84, 20, 22.9, 18.3, 18.9, 17.4, 17.6,  
#>     18, 17.98, 17.82, 17.42, 19.47, 18.52, 19.9, 20.01, 16.87,  
#>     17.3, 15.41, 17.05, 18.9, 16.7, 16.9, 14.5, 15.5, 14.6, 18.6 
#>     ), vs = c(0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0,  
#>     0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1), am = c(1,  
#>     1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1,  
#>     0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1), gear = c(4, 4, 4, 3,  
#>     3, 3, 3, 4, 4, 4, 4, 3, 3, 3, 3, 3, 3, 4, 4, 4, 3, 3, 3,  
#>     3, 3, 4, 5, 5, 5, 5, 5, 4), carb = c(4, 4, 1, 1, 2, 1, 4,  
#>     2, 2, 4, 4, 3, 3, 3, 4, 4, 4, 1, 2, 1, 1, 2, 2, 4, 2, 1,  
#>     2, 2, 4, 6, 8, 2)), row.names = c("Mazda RX4", "Mazda RX4 Wag",  
#> "Datsun 710", "Hornet 4 Drive", "Hornet Sportabout", "Valiant",  
#> "Duster 360", "Merc 240D", "Merc 230", "Merc 280", "Merc 280C",  
#> "Merc 450SE", "Merc 450SL", "Merc 450SLC", "Cadillac Fleetwood",  
#> "Lincoln Continental", "Chrysler Imperial", "Fiat 128", "Honda Civic",  
#> "Toyota Corolla", "Toyota Corona", "Dodge Challenger", "AMC Javelin",  
#> "Camaro Z28", "Pontiac Firebird", "Fiat X1-9", "Porsche 914-2",  
#> "Lotus Europa", "Ford Pantera L", "Ferrari Dino", "Maserati Bora",  
#> "Volvo 142E"), class = "data.frame")
#>      AIC      BIC   logLik deviance df.resid 
#>  35.2503  39.6475 -14.6251  29.2503       29 
#> Random effects:
#>  Groups Name        Std.Dev.
#>  vs     (Intercept) 0.7896  
#> Number of obs: 32, groups:  vs, 2
#> Fixed Effects:
#> (Intercept)          mpg  
#>     -8.7018       0.4085  

construct_model(
  x = mtcars |> dplyr::rename(`M P G` = mpg),
  formula = reformulate2(c("M P G", "cyl"), response = "hp"),
  method = "lm"
) |>
  ard_regression() |>
  dplyr::filter(stat_name %in% c("term", "estimate", "p.value"))
#> {cards} data frame: 6 x 7
#>   variable variable_level   context stat_name stat_label    stat
#> 1    M P G             NA regressi…      term       term `M P G`
#> 2    M P G             NA regressi…  estimate  Coeffici…  -2.775
#> 3    M P G             NA regressi…   p.value    p-value   0.213
#> 4      cyl             NA regressi…      term       term     cyl
#> 5      cyl             NA regressi…  estimate  Coeffici…  23.979
#> 6      cyl             NA regressi…   p.value    p-value   0.003
#> ℹ 1 more variable: fmt_fn